+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v 2.5.0 (?)
 Changes in this release:
+- Fix bug where the temporary directory used to store the Inmanta project is not cleaned up when an exception occurs in the setup stage of the project\_factory fixture.
 
 # v 2.4.0 (2022-09-07)
 Changes in this release: